+++ src/entry.py
@@ -1,7 +1,7 @@
 import sys
-from controller.singleplayer_controller import Singleplayer_Controller
+from controller.singleplayerController import SingleplayerController
 
 if __name__ == '__main__':
-    gameplay_controller = Singleplayer_Controller()
+    gameplay_controller = SingleplayerController()
     gameplay_controller.new_game()
     gameplay_controller.run_game()
